+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/BacklinksDelegateImpl.kt
@@ -74,7 +74,8 @@ internal class BacklinksDelegateImpl<T : TypedRealmObject>(
         reference: EmbeddedRealmObject,
         targetProperty: KProperty<*>
     ): T = createBacklinks(reference, targetProperty).firstOrNull()
-        ?: throw IllegalStateException("No object matches backlink '${targetProperty.name}'")
+//        ?: throw IllegalStateException("No object matches backlink '${targetProperty.name}'")
+        ?: throw IllegalStateException("Backlink '${targetProperty.name}' is not an instance of target property type '${sourceClass!!.simpleName}'.")
 
     override fun getValue(
         reference: RealmObject,
